pyproject.toml
[project]
name = "concrete-crack-detection"
version = "0.1.0"
description = "Automated crack detection system using deep learning for concrete surface classification"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "MIT" }

dependencies = [
    "tensorflow>=2.15.0,<2.16.0",
    "numpy>=1.26.3,<2.0.0",
    "pandas>=2.1.4,<3.0.0",
    "matplotlib>=3.8.2,<4.0.0",
    "seaborn>=0.13.1,<0.14.0",
    "pillow>=10.2.0,<11.0.0",
    "scikit-learn>=1.4.0,<2.0.0",
    "jupyter>=1.0.0,<2.0.0",
    "ipykernel>=6.28.0,<7.0.0",
    "kaggle>=1.5.16,<2.0.0",
    "flask>=3.0.0,<4.0.0",
    "requests>=2.31.0,<3.0.0",
]

[project.optional-dependencies]
# Production dependencies (for deployment)
prod = [
    "flask>=3.0.0,<4.0.0",
    "gunicorn>=21.2.0,<22.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "requests>=2.31.0,<3.0.0",
]

[dependency-groups]
# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
